Move data related code to same directory
