+++ dockerfile/elasticdl.dev
@@ -13,17 +13,21 @@ RUN pip install grpcio-tools
 # Install the Kubernetes Python client
 RUN pip install kubernetes
 
-# Install RecordIO and its dependency
-RUN apt-get install -y libsnappy-dev && \
-    pip install pyrecordio docker
+# Install Docker python SDK
+RUN pip install docker 
+
+# Install RecordIO dependencies
+RUN apt-get install -y libsnappy-dev
+
+# Install RecordIO 
+RUN pip install 'pyrecordio>=0.0.2'
 
 ENV PYTHONPATH=/
 
-# RecordIO dataset generator
-COPY recordio_ds_gen /recordio_ds_gen  
-COPY record_codec /record_codec
-RUN python /recordio_ds_gen/mnist/gen_data.py --codec-type tf_example /data
-RUN python /recordio_ds_gen/cifar10/gen_data.py /data
+# ElasticDL dataset generator and codecs
+COPY edl_data /edl_data  
+RUN python /edl_data/recordio_gen/mnist/gen_data.py --codec-type tf_example /data
+RUN python /edl_data/recordio_gen/cifar10/gen_data.py /data
 
 WORKDIR /
 COPY elasticdl /elasticdl
