+++ edl_data/codec/__init__.py
@@ -1,0 +1,4 @@
+from edl_data.codec.bytes_codec import BytesCodec
+from edl_data.codec.tf_example_codec import TFExampleCodec
+
+__all__=[BytesCodec, TFExampleCodec]
