+++ edl_data/codec/bytes_codec.py
