+++ edl_data/recordio_gen/cifar10/record.py
