+++ edl_data/recordio_gen/mnist/show_data.py
