--- record_codec/__init__.py
@@ -1,4 +1,0 @@
-from record_codec.bytes_codec import BytesCodec
-from record_codec.tf_example_codec import TFExampleCodec
-
-__all__=[BytesCodec, TFExampleCodec]
